[MR] | 2020.08.23 | 'front' into 'develop'

+++ backend/src/main/java/com/web/blog/service/CommentService.java
@@ -22,15 +22,16 @@ public class CommentService {
 
 	private final CommentRepository commentRepository;
 	private final UsersRepository userRepository;
+
+	private final String ADMIN = "ROLE_ADMIN";
 	
 	public void createComments(Comments comment) {
-		System.out.println(comment);
 		commentRepository.save(Comments.builder()
 				.pid(comment.getPid())
 				.comment(comment.getComment())
 				.email(comment.getEmail())
-				.comment_time(LocalDateTime.now())
-				.update_time(LocalDateTime.now())
+				.commentTime(LocalDateTime.now())
+				.updateTime(LocalDateTime.now())
 				.build());
 		// 댓글 작성 시 댓글 작성자 경험치 상승
 		Optional<Users> user = userRepository.findByEmail(comment.getEmail());
@@ -42,7 +43,7 @@ public class CommentService {
 	
 	public List<Comments> listAllComments(int pid){
 		List<Comments> result = new ArrayList<Comments>();
-		result = commentRepository.findByPid(pid);
+		result = commentRepository.findAllByPidOrderByCommentTimeDesc(pid);
 		return result;
 	}
 	
@@ -59,8 +60,14 @@ public class CommentService {
 	}
 	
 	@Transactional
-	public void deleteComments(int cmid) {
-		commentRepository.deleteByCmid(cmid);
+	public boolean deleteComments(String user, int cmid, String role) {
+		Comments comment = commentRepository.findByCmid(cmid);
+		if(user.equals(comment.getEmail()) | role.equals(ADMIN)) {
+			commentRepository.deleteByCmid(cmid);
+			return true;
+		}else {
+			return false;
+		}
 	}
 
 }
