+++ backend/src/main/java/com/web/blog/service/PostService.java
@@ -33,9 +33,9 @@ public class PostService {
 	private final MemberRepository memberRepository;
 	private final UsersRepository userRepository;
 
+	private final String ADMIN = "ROLE_ADMIN";
 	
 	public void createPost(Post post) {
-		System.out.println(post);
 		postRepository.save(Post.builder()
 				.bid(post.getBid())
 				.lcid(post.getLcid())
@@ -45,7 +45,7 @@ public class PostService {
 				.author(post.getAuthor())
 				.manager(post.getManager())
 				.postTime(LocalDateTime.now())
-				.update_time(LocalDateTime.now())
+				.updateTime(LocalDateTime.now())
 				.ptype(post.getPtype())
 				.build()).getPid();
 		// 게시글 작성 시 작성자 경험치 상승
@@ -56,20 +56,7 @@ public class PostService {
 		});
 	}
 	
-//	public boolean countBlogByUser(String email) {
-//		if(blogRepository.countByManager(email)>6) {
-//			return false;
-//		}else {
-//			return true;
-//		}
-//	}
-//	
 	public Page<Post> listAllPost(int bid, Pageable pageable){
-//		List<Post> result = new ArrayList<Post>();
-//		Page<Post> result = new ArrayList<Post>();
-//		result = postRepository.findAllByBid(bid);
-		// 최신글 순서로 조회 
-//		result = postRepository.findAllByBidAndPtypeIsNullOrderByPostTimeDesc(bid);
 		Page<Post> result = postRepository.findAllByBidAndPtypeIsNullOrderByPostTimeDesc(bid, pageable);
 		return result;
 	}
@@ -97,22 +84,19 @@ public class PostService {
 		return post;
 	}
 	
-	public void updatePost(Post post,String bucketName, String accessKey, String secretKey) {
+	public void updatePost(String content, Post post,String bucketName, String accessKey, String secretKey) {
 		Post updatePost = postRepository.findByPid(post.getPid());
 		
-		String content = updatePost.getPcontent();
 		System.out.println("[이전 데이터] "+content);
 		System.out.println("[바뀔 데이터] "+post.getPcontent());
 		
-		if(content.contains("img")) {
+		if(content.contains("img") && !content.equals(post.getPcontent())) {
 			String[] inputArr = content.split("'");
 			for(int i=0;i<inputArr.length;i++) {
 				if(inputArr[i].contains("https://memody")) {
 					String[] temp = inputArr[i].split("/");
 					S3Util s3 = new S3Util(accessKey, secretKey);
 					String temp2 = temp[3]+"/"+temp[4]+"/"+temp[5]+"/"+temp[6]+"/"+temp[7];
-					System.out.println(temp2);
-					System.out.println(temp[temp.length-1]);
 					s3.fileDelete(bucketName, temp2);
 				}
 			}
@@ -126,14 +110,18 @@ public class PostService {
 	}
 	
 	@Transactional
-	public void deletePost(int pid) {
-		postRepository.deleteByPid(pid);
+	public boolean deletePost(String user, int pid, String role) {
+		Post post = postRepository.findByPid(pid);
+		if(user.equals(post.getManager()) | role.equals(ADMIN)) {
+			postRepository.deleteByPid(pid);
+			return true;
+		}else {
+			return false;
+		}
 	}
 
 	public Page<Post> listAllPostByMCategory(int bid, int mcid, Pageable pageable){
-//		List<Post> result = new ArrayList<Post>();
 		// 최신글 순서로 조회 
-//		result = postRepository.findAllByBidAndMcidAndPtypeIsNullOrderByPostTimeDesc(bid, mcid);
 		Page<Post> result = postRepository.findAllByBidAndMcidAndPtypeIsNullOrderByPostTimeDesc(bid, mcid, pageable);
 		return result;
 	}
@@ -176,7 +164,7 @@ public class PostService {
 				.author(post2.getAuthor())
 				.manager(user)
 				.postTime(LocalDateTime.now())
-				.update_time(LocalDateTime.now())
+				.updateTime(LocalDateTime.now())
 				.ptype(post2.getPtype())
 				.build());
